config: create global config file
