+++ .circleci/config.yml
@@ -3,7 +3,7 @@ jobs:
   build_stg:
     docker:
       - image: circleci/node:8.10
-    working_directory: ~/escritorio-pagarme
+    working_directory: ~/the-office
     steps:
       - checkout
       - restore_cache:
@@ -34,7 +34,7 @@ jobs:
   build_prd:
     docker:
       - image: circleci/node:8.10
-    working_directory: ~/escritorio-pagarme
+    working_directory: ~/the-office
     steps:
       - checkout
       - restore_cache:
